New CreateUnsignablePdf tests focusing on the stack overflow question
why form cannot be signed (error message “due to current status”)
https://stackoverflow.com/questions/56393909/why-form-cannot-be-signed-error-message-due-to-current-status

testCreateLikeJS - Indeed, the PDF created by the OP triggers a "The document cannot be signed in its current state. Please save the document, close it, reopen it, and then attempt to sign again." The "current state" might be one where JavaScript replaced event handlers. In the document at hand, for example, the domicilation event handler defines a function fft() and sets it as the new event handler. But this causes the definition of the function fft() to be only present in the PDF viewer memory and nowhere in the PDF anymore. A signed (and in the process of signing saved) PDF, therefore, would behave differently than the PDF present for signing.

testCreateLikeJSImproved - In contrast to the OP's original code the JavaScript code here defines the functions fft() and fft2() outside the original event handler. Thus, even after triggering the original event handlers the definition of these functions is still present in the PDF. Nonetheless this document also triggers a "The document cannot be signed in its current state. Please save the document, close it, reopen it, and then attempt to sign again." message. Probably Adobe Reader does not check whether all code in question is still defined somewhere in the PDF but automatically considers a PDF with manipulated event listeners dirty...