New ExtractText tests focusing on wen li's answer to the stack overflow question
PDFBox extracting paragraphs
https://stackoverflow.com/a/56580253/1729265
PDF32000_2008.pdf
https://www.adobe.com/content/dam/acom/en/devnet/pdf/pdfs/PDF32000_2008.pdf

testPDF32000pageii - Here it looks the other way around compared to what the OP claims: there is a space at the end of all but the last paragraph line.

testPDF32000pagevii - Here one sees that there is not always a space at the end of all the non-last paragraph lines, "PDF/X" is split as "PDF/" and "X" between lines, and there is no space in-between.