New FillInForm tests focusing on the stack overflow question
PDFBox Inconsistent PDTextField Autosize Behavior after setValue
https://stackoverflow.com/questions/56938135/pdfbox-inconsistent-pdtextfield-autosize-behavior-after-setvalue
0.pdf
http://www.filedropper.com/0postfontload

testFill0LikeXenyal - Indeed, some fields look weird after fill-in; for some fields this is due to weird pre-existing appearance streams.

testFill0DropOldAppearance - Removing the old appearance streams before setting the new field values removes the compression of the Resident Name and Care Providers Address fields. In the latter case, though, the lower part of the field value now is cut off. For some fields only the first two letters are visible. This is due to them being two character comb fields.

testFill0DropOldAppearanceNoCombNoMax - Resetting the comb flags and removing maximum lengths fixes the appearance of the fields in which only the first two letters were visible. The problem of the lower part of the field value being cut off in the Care Providers Address fields is due to PDFBox in case of  multi line text fields using a fixed font height and not fine tuning the vertical position of the field contents.

testFill0DropOldAppearanceNoCombNoMaxNoMultiLine - By resetting the MultiLine flags, too, one eventually gets rid of the problem of the lower part of the field value being cut off in the Care Providers Address fields. This actually should be considered an issue of PDFBox, though, not of the source PDF here.