New AddSpecialCharacterWithoutEmbedding test focusing on the stack overflow question
Without embeded fonts, is PDF limited to only 4281 characters (of AGL)? How to display more glyphs?
https://stackoverflow.com/questions/57295834/without-embeded-fonts-is-pdf-limited-to-only-4281-characters-of-agl-how-to-d

testWithArial - This test constructs a type 1 font without embedded font program that uses a non AGL name in its encoding's Differences array to provide a glyph not to be found on the AGL. This is a counterexample to the OP's original question.