New CreateSignature tests focusing on the stack overflow question
How to apply digital signature image at bottom left position in the last
page of pdf using pdfbox?
https://stackoverflow.com/questions/58427451/how-to-apply-digital-signature-image-at-bottom-left-position-in-the-last-page-of
pdf-test.pdf
http://www.orimi.com/pdf-test.pdf

As the OP found out himself, the `BoundingBoxFinder` coordinates could
not be used as is in the `CreateVisibleSignature`.

signLikeHemantPdfTest - This test demonstrates the required
transformation with the example document apparently used by the OP.

signLikeHemantTest - This test demonstrates the required transformation
with an arbitrary example document.