New MakeLtvEnabled test focusing on the stack overflow question
Get tag's related BBox's even though there is no attributes (/A in
document catalog structure) related to Layout in PDFBox?
https://stackoverflow.com/questions/59192443/get-tags-related-bboxs-even-though-there-is-no-attributes-a-in-document-cata
res_multipage.pdf
https://drive.google.com/file/d/1_-tuWuReaTvrDsqQwldTnPYrMHSpXIWp/view?usp=sharing

testVisualizeResMultipage - This test shows how to determine the
bounding box of text content in structure elements for the document in
question.

testVisualizeTestWPhromma - This test shows how to determine the
bounding box of text content in structure elements for a previously
added document.

visualize plus helpers - These methods output an XML'ish representation
of the structure tree plus text extracted for it and additionally
creates a PDF with frames representing the bounding boxes of the text
inside the structure elements.