New OpenFile test focusing on the stack overflow question
Taking screenshots from PDF file with Apache PDFBox
https://stackoverflow.com/questions/59604736/taking-screenshots-from-pdf-file-with-apache-pdfbox
samplePDF.pdf
http://aplaidshirt.epizy.com/samplePDF.pdf

According to the stack trace the error occurred during PDDocument.load.
Thus, this test tries to load the example file. Unfortunately, no error
happens. Thus, this might be an issue in the OP's version of PDFBox
which meanwhile is fixed.