New EditPageContent test focusing on the stack overflow question
Extract content stream(Images, Text and graphics) with in a BBOX. And
place it back in new PDF without loosing any style?
https://stackoverflow.com/questions/59489717/extract-content-streamimages-text-and-graphics-with-in-a-bbox-and-place-it-b
Cengage1.pdf
https://drive.google.com/file/d/1q8e3475Urg56fo5svuckjlbqxspElQ2s/view?usp=sharing

testSortDrawsCengage1 - This test tries to sort text drawing
instructions by area. It is a mere proof of concept, though, even less
actually, as it creates an invalid marked content structure (multiple
marked content sections with the same ID.
This allowed, though, to identify a bug in PdfContentStreamEditor - a
few OperatorProcessor implementations recursively feed other operators
by which they are defined into the stream engine. For the purpose of
editing one has to ignore those recursive processing calls which
PdfContentStreamEditor did not do yet. This is fixed now.