Fix PdfContentStreamEditor
In the EditPageContent test testSortDrawsCengage1 a bug in
PdfContentStreamEditor was identified - a few OperatorProcessor
implementations recursively feed other operators by which they are
defined into the stream engine. For the purpose of editing one has to
ignore those recursive processing calls which PdfContentStreamEditor did
not do yet. This is fixed now.