New DenseMerging test focusing on the stack overflow question
How to dense merge PDF files using PDFBox 2 without whitespace near page
breaks?
https://stackoverflow.com/questions/60052967/how-to-dense-merge-pdf-files-using-pdfbox-2-without-whitespace-near-page-breaks

testVerticalAnalyzer - This test checks the PageVerticalAnalyzer
functionality. Beware, as mentioned in the PageVerticalAnalyzer
comments, the processing in particular of curves is incorrect. The curve
used in this test is chosen not to create wrong results due to this
known issue.

PageVerticalAnalyzer - This PDFGraphicsStreamEngine looks for vertical
sections of use in a page. Beware, this is a mere proof of concept.
Especially curveTo needs to be improved, see the comment there. This
class corresponds to the itext5 test area class with the same name,
mkl.testarea.itext5.merge.PageVerticalAnalyzer.