New CurvedBorderRectangle tests focusing on the stack overflow question
Can't Get Path to Fill in PDFBox
https://stackoverflow.com/questions/60942642/cant-get-path-to-fill-in-pdfbox

testLikeMaht33n - The reason for this code not creating a filled
rectangle is that the path contains many moveTo instructions effectively
splitting the path into a multitude of subpaths each of which is filled
by itself but none is encompassing the whole rectangle.

testLikeMaht33nImproved - This improves the OP's code from
testLikeMaht33n by drawing the whole rectangle outline in one stroke
with no moveTo but the initial one. This required changing directions of
some lines and curves. Now filling the path results in the desired
filled rectangle.