New EditPageContent test focusing on the stack overflow question
Replace or remove text from PDF with PDFbox in Java
https://stackoverflow.com/questions/63592078/replace-or-remove-text-from-pdf-with-pdfbox-in-java
nuevo.pdf
http://www.mediafire.com/file/9w3kkc4yozwsfms/file

testRemoveQrTextNuevo - This test shows how to look for specific text and remove it. Beware, this is a simple case where the exactly the whole search text is drawn by a single text drawing instruction, so the code only has to look at the most recently drawn characters. In general one may have to collect some text pieces and all instructions executed at that time and only forward them to super or drop some of them if the search term is found. In the worst case the whole page has to be collected, the text pieces sorted, and from that set of sorted texts instructions to keep or not to keep have to be determined.